requirements.txt
polars
psycopg2-binary
python-dotenv
typing-extensions
numpy